requirements.txt
matplotlib==3.8.1
tensorflow==2.13.1
scikeras==0.12.0
scikit-learn==1.3.2
